seed fixer and db bigint

+++ db/knex_migrations/20180206004502_initializeDB.js
@@ -8,7 +8,7 @@ exports.up = function (knex, Promise) {
             table.increments();
             table.string(booksTable.title).notNullable();
             table.string(booksTable.description, 2040).notNullable();
-            table.integer(booksTable.ISBN).notNullable();
+            table.bigInteger(booksTable.ISBN).notNullable();
             table.string(booksTable.author).notNullable();
             table.date(booksTable.publication_date).notNullable();
             table.enu(booksTable.genre, GENERES).notNullable();
